Fix: corregir tipos double en obtenerPrecioMasBarato()
- Cambiar int por double en testPrecioMasBaratoEsValido()
- Cambiar int por double en testCompararPreciosEntreBusquedas()
- Actualizar assertNotEquals de -1 a -1.0 para compatibilidad

+++ src/test/java/pe/meli/tests/selenium/Ejercicio5_POMTest.java
@@ -40,10 +40,10 @@ class Ejercicio5_POMTest extends BaseTest {
 
         searchPage.buscar("teclado");
 
-        int precioMasBarato = searchPage.obtenerPrecioMasBarato();
+        double precioMasBarato = searchPage.obtenerPrecioMasBarato();
 
         System.out.println(">>> Precio más barato encontrado: S/ " + precioMasBarato);
-        assertNotEquals(-1, precioMasBarato,
+        assertNotEquals(-1.0, precioMasBarato,
                 "No se encontró ningún precio en la página");
         assertTrue(precioMasBarato > 0,
                 "El precio debe ser mayor que 0, se obtuvo: " + precioMasBarato);
